Fixed Mono Issue, Added different way of triggering emitter
I fixed the mono issue by removing the SkullAmb_Trigger event's associated audio file in wwise, and generated the soundbank. Then I re-added the sound file/regenerate sound bank.

Unity still only updates wwise about the player distance to an object only once (on initial trigger detection I think).

To experiment with this, I added a different way of triggering the wwise events with the Sound Debug Cue (just walk past it to trigger the playback). I also added brand new emitter events in wwise to see if our issue was the event or not. It seems like it's an issue with Unity updating wwise.